requirements.txt
# Python 3.9 최적화 버전
torch==2.0.1
torchvision==0.15.2
numpy==1.24.3
opencv-python==4.8.0.74
Pillow==10.0.0

# Face detection (Python 3.9 완벽 호환)
facenet-pytorch==2.5.3
insightface==0.7.3
onnxruntime-gpu==1.15.1  # Python 3.9에서 안정적

# Image processing
albumentations==1.3.1
scikit-image==0.21.0

# Utilities
tqdm==4.65.0
pyyaml==6.0
loguru==0.7.0
pandas==2.0.3
seaborn==0.13.2